+++ src/main/java/org/sonar/plugins/delphi/antlr/ast/node/UnitImportNode.java
@@ -1,7 +1,9 @@
 package org.sonar.plugins.delphi.antlr.ast.node;
 
+import net.sourceforge.pmd.lang.symboltable.NameDeclaration;
 import org.antlr.runtime.Token;
 import org.sonar.plugins.delphi.antlr.ast.visitors.DelphiParserVisitor;
+import org.sonar.plugins.delphi.symbol.declaration.UnitImportNameDeclaration;
 
 public final class UnitImportNode extends DelphiNode {
   public UnitImportNode(Token token) {
@@ -19,5 +21,15 @@ public final class UnitImportNode extends DelphiNode {
 
   public QualifiedNameDeclarationNode getNameNode() {
     return (QualifiedNameDeclarationNode) jjtGetChild(0);
+  }
+
+  public boolean isResolvedImport() {
+    NameDeclaration declaration = getNameNode().getNameDeclaration();
+    return declaration != null
+        && ((UnitImportNameDeclaration) declaration).getOriginalDeclaration() != null;
+  }
+
+  public UnitImportNameDeclaration getImportNameDeclaration() {
+    return (UnitImportNameDeclaration) getNameNode().getNameDeclaration();
   }
 }